React env build

+++ frontend/src/hooks/users/useDeactivateAccountButtonAction.tsx
@@ -9,12 +9,17 @@ export const useDeactivateAccountButtonAction = () => {
   const [currentUser, setCurrentUser] = useRecoilState(currentUserState);
   const navigate = useNavigate();
   const onCLickDeactivateAccountAction: React.MouseEventHandler<HTMLButtonElement> = (e) => {
-    if(window.confirm('本当に退会しますか？')){
-      client.delete(`users/${currentUser.user.id}`).then(response => {
-        response.data.status === 'deleted' && navigate('/', {state: {message: '退会しました', type: 'success-message', condition: true}} );
-        response.data.status === 'deleted' && setLoginStatus({status: false}) && setCurrentUser({user:{}});
-      });
-    };
+    // guestユーザーは退会できなくする。
+    if( currentUser.user.id === 1 ){
+      navigate(`./`, {state: {message: 'guestユーザーは退会できません', type: 'error-message', condition: true}})
+    }else {
+      if(window.confirm('本当に退会しますか？')){
+        client.delete(`users/${currentUser.user.id}`).then(response => {
+          response.data.status === 'deleted' && navigate('/', {state: {message: '退会しました', type: 'success-message', condition: true}} );
+          response.data.status === 'deleted' && setLoginStatus({status: false}) && setCurrentUser({user:{}});
+        });
+      };
+    }
   };
 
   return { onCLickDeactivateAccountAction, loginStatus };
